Naive extraction of review reply module
